Hosting service mining process refactored.
